pyproject.toml
[tool.poetry]
name = "alexofechoes_gendiff"
version = "0.1.0"
description = "Gendiff - hexlet second project"
authors = ["Alexander Bukharin <alexawm182@gmail.com>"]
packages = [
    { include = "gendiff" },
]
repository = "https://github.com/alexofechoes/python-project-lvl2"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.1"
wemake-python-styleguide = "^0.12.4"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
flake8 = "^3.7"
pytest-cov = "^2.7"

[tool.poetry.scripts]
gendiff = "gendiff.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
